split testutil/loopback_linux.go to another pkg
containerd/containerd#2694 is going to import both `containerd/pkg/testutil` and
`continuity/testutil` for deduplicating `NewLoopback()`.

However, as the both pkg defines `test.root` BoolVar in `init()`, the test panics.

This commit avoids the panic issue by splitting `continuity/testutil/loopback_linux.go`
to another pkg (`continuity/testutil/loopback`), so taht containerd does not need to import
both `containerd/pkg/testutil` and `continuity/testutil`.
i.e. containerd will import `containerd/pkg/testutil` and `continuity/testutil/loopback`.
